% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{leaf_class_colors}
\alias{leaf_class_colors}
\title{Default class colours}
\usage{
leaf_class_colors(multiclass = FALSE)
}
\arguments{
\item{multiclass}{Logical.}
}
\value{
Named character vector of hex colours for non-background classes.
}
\description{
Default class colours
}
\examples{
leaf_class_colors(TRUE)
}
