% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_io.R
\name{leaf_image_metadata}
\alias{leaf_image_metadata}
\title{Metadata of a leaf image}
\usage{
leaf_image_metadata(image)
}
\arguments{
\item{image}{An image returned by [read_leaf()] or any EBImage `Image`.}
}
\value{
A list with `filename`, `extension`, `width`, `height`, `channels`
  (`NA` where unknown).
}
\description{
Metadata of a leaf image
}
\examples{
leaf_image_metadata(read_leaf(leaf_example_images()[2]))
}
