% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_io.R
\name{list_leaf_images}
\alias{list_leaf_images}
\title{List image files in a directory}
\usage{
list_leaf_images(dir, recursive = FALSE)
}
\arguments{
\item{dir}{Directory.}

\item{recursive}{Search sub-directories.}
}
\value{
Character vector of image paths with supported extensions.
}
\description{
List image files in a directory
}
\examples{
list_leaf_images(system.file("extdata", package = "LeafInjuryR"))
}
