% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{mask_to_image}
\alias{mask_to_image}
\title{Render a logical mask as a black/white image}
\usage{
mask_to_image(mask)
}
\arguments{
\item{mask}{Logical matrix.}
}
\value{
A grayscale `Image` (leaf = white).
}
\description{
Render a logical mask as a black/white image
}
\examples{
mask_to_image(matrix(c(TRUE, FALSE), 4, 4))
}
