% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_io.R
\name{write_image_png}
\alias{write_image_png}
\title{Write a colour image to PNG}
\usage{
write_image_png(image, path)
}
\arguments{
\item{image}{Colour `Image` or array `[width, height, 3]`.}

\item{path}{Output file path.}
}
\value{
`path`, invisibly.
}
\description{
Write a colour image to PNG
}
\examples{
f <- tempfile(fileext = ".png")
write_image_png(make_synthetic_leaf()$image, f)
}
