% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write_leaf_results}
\alias{write_leaf_results}
\title{Write tabular results to CSV}
\usage{
write_leaf_results(x, file)
}
\arguments{
\item{x}{A `leaf_analysis`, `leaf_batch` or data frame of metrics.}

\item{file}{Output CSV path.}
}
\value{
`file`, invisibly.
}
\description{
Write tabular results to CSV
}
\examples{
res <- analyze_leaf(leaf_example_images()[1])
write_leaf_results(res, tempfile(fileext = ".csv"))
}
